3A-G05-W003. Per-File Export of Interfaces and turn the ToastProps to ToastParams
