+++ app/admin/page.tsx
@@ -33,8 +33,8 @@ import {
 import { useAuth } from "@/lib/contexts/AuthContext";
 import { useDeleteRequest } from "@/lib/contexts/DeleteRequestContext";
 import { StatCard } from "@/components/dashboard/StatCard";
-import { ToastContainer, type ToastProps } from "@/components/ui/Toast";
-import { mockAdminUsers, type AdminUser } from "@/lib/data/mockData";
+import { ToastContainer, type ToastParams } from "@/components/ui/Toast";
+import { mockAdminUsers, type AdminUser } from "@/lib/data/data";
 import { generateId } from "@/lib/utils/formatters";
 
 // ─── Types ────────────────────────────────────────────────────────────────────
@@ -96,7 +96,7 @@ export default function AdminDashboardPage() {
 
   // Local state
   const [users, setUsers] = useState<AdminUser[]>(mockAdminUsers);
-  const [toasts, setToasts] = useState<Omit<ToastProps, "onClose">[]>([]);
+  const [toasts, setToasts] = useState<Omit<ToastParams, "onClose">[]>([]);
   const [search, setSearch] = useState("");
   const [sortKey, setSortKey] = useState<SortKey>("name");
   const [sortDir, setSortDir] = useState<SortDir>("asc");
@@ -108,7 +108,7 @@ export default function AdminDashboardPage() {
     return () => clearTimeout(t);
   }, []);
 
-  function addToast(message: string, type: ToastProps["type"] = "info") {
+  function addToast(message: string, type: ToastParams["type"] = "info") {
     setToasts((prev) => [...prev, { id: generateId(), message, type }]);
   }
 
@@ -364,7 +364,7 @@ export default function AdminDashboardPage() {
                   />
                 </PieChart>
               </ResponsiveContainer>
-              <div className="absolute inset-0 h-[200px] flex flex-col items-center justify-center pointer-events-none">
+              <div className="absolute inset-0 h-50 flex flex-col items-center justify-center pointer-events-none">
                 <span className="font-display font-bold text-3xl text-litter-text leading-none">
                   {agg.totalCats}
                 </span>
@@ -578,7 +578,7 @@ export default function AdminDashboardPage() {
           </div>
 
           <div className="overflow-x-auto">
-            <table className="w-full min-w-[680px] text-sm">
+            <table className="w-full min-w-170 text-sm">
               <thead>
                 <tr className="bg-litter-bg border-b border-litter-border">
                   <th className="text-left px-6 py-3">
