+++ components/ui/Toast.tsx
@@ -3,14 +3,7 @@
 import { useEffect } from "react";
 import { motion, AnimatePresence } from "framer-motion";
 import { X, CheckCircle, AlertCircle, Info } from "lucide-react";
-
-export interface ToastProps {
-  id: string;
-  message: string;
-  type?: "success" | "error" | "info" | "warning";
-  duration?: number;
-  onClose: (id: string) => void;
-}
+import { ToastParams } from "@/lib/interfaces/toastprops";
 
 export function Toast({
   id,
@@ -18,7 +11,7 @@ export function Toast({
   type = "info",
   duration = 3000,
   onClose,
-}: ToastProps) {
+}: ToastParams) {
   useEffect(() => {
     const timer = setTimeout(() => {
       onClose(id);
@@ -87,3 +80,5 @@ export function ToastContainer({ toasts, onClose }: ToastContainerProps) {
     </div>
   );
 }
+export type { ToastParams };
+
