+++ lib/data/data.ts
@@ -9,76 +9,14 @@
  * - Device-level data: air quality, litter level (shared across all cats — one litter box)
  */
 
-export interface Cat {
-  id: string;
-  name: string;
-  status: "healthy" | "watch" | "alert";
-  avatar: string | null;
-  isOnline: boolean;
-}
+import { Cat } from "../interfaces/cat";
+import { CatStats} from "../interfaces/catStats";
+import { ActivityItemData } from "../interfaces/activityItemData";
+import { CatDetails } from "../interfaces/catDetails";
+import { Session } from "../interfaces/session";
+import { HealthLog } from "../interfaces/healthlog";
+import { PastReport } from "../interfaces/pastReport";
 
-export interface CatStats {
-  visits: number;
-  avgDuration: string;
-  airQuality: "Normal" | "Elevated" | "Poor";
-  litterLevel: number;
-  lastVisit: string;
-}
-
-export interface ActivityItemData {
-  catId: string;
-  action: string;
-  time: string;
-  duration?: string;
-  anomaly: boolean;
-  anomalyNote?: string;
-}
-
-export interface CatDetails {
-  breed: string;
-  dob: string;
-  weightKg: number;
-  rfidTag: string;
-  healthInsight: string;
-  baseline: {
-    avgVisitsPerDay: number;
-    avgDurationSecs: number;
-    mq135DeltaPercent: number;
-    mq136DeltaPercent: number;
-    lastUpdated: string;
-  };
-}
-
-export interface Session {
-  id: string;
-  catId: string;
-  date: string;
-  time: string;
-  durationSecs: number;
-  mq135Delta: number;
-  mq136Delta: number;
-  anomaly: boolean;
-  anomalyType: string | null;
-  sessionStatus?: string;
-  summaryVisits?: number;
-}
-
-export interface HealthLog {
-  id: string;
-  catId: string;
-  date: string;
-  type: "Vet Visit" | "Medication" | "Observation" | "Other";
-  note: string;
-}
-
-export interface PastReport {
-  id: string;
-  catId: string;
-  catName: string;
-  range: string;
-  generatedOn: string;
-  filename: string;
-}
 // Mock data for testing and development
 // ⚠️ EMPTY ARRAYS — Users add their own cats manually
 export const mockCats: Cat[] = [];
@@ -420,3 +358,4 @@ export function getTrendData(catId: string) {
     };
   });
 }
+export type { HealthLog, Session, PastReport};
