+++ lib/interfaces/activityItemData.ts
@@ -1,0 +1,8 @@
+export interface ActivityItemData {
+    catId: string;
+    action: string;
+    time: string;
+    duration?: string;
+    anomaly: boolean;
+    anomalyNote?: string;
+  }
