+++ lib/interfaces/cat.ts
@@ -1,0 +1,7 @@
+export interface Cat {
+    id: string;
+    name: string;
+    status: "healthy" | "watch" | "alert";
+    avatar: string | null;
+    isOnline: boolean;
+  }
