+++ lib/interfaces/catStats.ts
@@ -1,0 +1,7 @@
+export interface CatStats {
+  visits: number;
+  avgDuration: string;
+  airQuality: "Normal" | "Elevated" | "Poor";
+  litterLevel: number;
+  lastVisit: string;
+}
