+++ lib/interfaces/healthlog.ts
@@ -1,0 +1,7 @@
+export interface HealthLog {
+    id: string;
+    catId: string;
+    date: string;
+    type: "Vet Visit" | "Medication" | "Observation" | "Other";
+    note: string;
+  }
