+++ lib/interfaces/pastReport.ts
@@ -1,0 +1,8 @@
+export interface PastReport {
+    id: string;
+    catId: string;
+    catName: string;
+    range: string;
+    generatedOn: string;
+    filename: string;
+  }
